% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_soccer_logos.R, R/soccerplotR-package.R
\docType{data}
\name{GeomSoccerlogo}
\alias{GeomSoccerlogo}
\alias{soccerplotR}
\alias{soccerplotR-package}
\title{soccerplotR: Soccer Logo Plots in 'ggplot2'}
\format{
An object of class \code{GeomSoccerlogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomSoccerlogo
}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of functions to visualize professional soccer analysis in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tonyelhabr/soccerplotR}
  \item \url{https://tonyelhabr.github.io/soccerplotR/}
  \item Report bugs at \url{https://github.com/tonyelhabr/soccerplotR/issues}
}

}
\author{
\strong{Maintainer}: Tony ElHabr \email{mrcaseb@gmail.com}

}
\keyword{datasets}
\keyword{internal}
