% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team_name_mapping}
\alias{team_name_mapping}
\title{Alternate Team Abbreviation Mapping}
\format{
A list of named character vectors, where each element in the list
is an array of teams corresponding to a given country (3-letter ISO code).
Each character vector has the following form:
\describe{
\item{name attribute}{The "alternate" name.}
\item{value attribute}{The "correct" name.}
}
}
\usage{
team_name_mapping
}
\description{
A list of named character vectors mapping common alternate team abbreviations.
}
\details{
You can suggest additions to this table by
\href{https://github.com/tonyelhabr/soccerplotR/issues/new}{opening an issue in soccerplotR}.
}
\examples{
\donttest{
library(purrr)

available_countries <- names(soccerplotR::team_name_mapping)
available_countries
#> [1] "ENG" "ESP" "FRA" "GER" "ITA" "USA"

eng_teams <- soccerplotR::team_name_mapping[["ENG"]]
eng_teams[c("Wolves", "Wolverhampton Wanderers", "WBA", "West Brom")]

all_teams <- flatten_chr(soccerplotR::team_name_mapping)
all_teams[c("Tottenham Hotspur", "Tottenham", "AC Milan")]
}

}
\keyword{datasets}
