% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_names}
\alias{valid_names}
\alias{valid_team_names}
\alias{all_valid_team_names}
\title{Output Valid Soccer Team Names}
\usage{
valid_team_names(country)

all_valid_team_names()
}
\arguments{
\item{country}{One of "ENG", "ESP", "FRA", "GER", "ITA", or "USA". (Names are ISO 3166 3-letter country codes.)}
}
\value{
A vector of type \code{'character'}.
}
\description{
Team names used in this function are extracted from Fotmob
}
\examples{
all_valid_team_names()
valid_team_names("ENG")
}
