% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_team_names}
\alias{valid_team_names}
\title{Output Valid Soccer Team Abbreviations}
\usage{
valid_team_names()
}
\value{
A named vector of type \code{'character'}. Names are ISO 3166 3-letter country codes
}
\description{
Thenames used in this function are extracted from Fotmob
}
\examples{
valid_team_names()
}
