% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_team_names}
\alias{clean_team_names}
\title{Standardize Soccer Team Names}
\usage{
clean_team_names(team_name, keep_non_matches = TRUE)
}
\arguments{
\item{keep_non_matches}{If \code{TRUE} (the default) an element of \code{team_name} that can't
be matched to any of the internal mapping vectors will be kept as is. Otherwise
it will be replaced with \code{NA}.}
}
\value{
A character vector with the length of \code{team_name} and cleaned team names
if they are included in \code{\link{team_name_mapping}}, etc.
(depending on the value of \code{country}). Non matches may be replaced
with \code{NA} (depending on the value of \code{keep_non_matches}).
}
\description{
This function standardizes soccer team names to Fotmob defaults.
}
\examples{
team_names <- c('Liverpool', 'Brighton', 'Bournemouth', 'AFC Bournemouth')

# keep non matches
soccerplotR::clean_team_names(team_names, keep_non_matches = TRUE)
}
