% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_team_names}
\alias{valid_team_names}
\title{Output Valid Soccer Team Abbreviations}
\usage{
valid_team_names(country = c("ENG", "USA"))
}
\arguments{
\item{country}{One of \code{'ENG'} or \code{'USA'}}
}
\value{
A vector of type \code{'character'}.
}
\description{
Thenames used in this function are extracted from Fotmob
}
\examples{
valid_team_names('ENG')
valid_team_names('USA')
}
